Move assignLyrics to LyricService
